% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{bar_charts_fr}
\alias{bar_charts_fr}
\title{Lance bar_charts_fr}
\usage{
bar_charts_fr(file = tutor:::tous_les_programmes()[1], ...)
}
\arguments{
\item{file}{name of the learnr to launch}

\item{...}{other params}
}
\description{
Lance bar_charts_fr
}
