% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions_en.R
\name{boxplots_en}
\alias{boxplots_en}
\title{launch boxplots_en}
\usage{
boxplots_en(file = tutor:::tous_les_programmes(lang = "en")[11], ...)
}
\arguments{
\item{file}{name of the learnr to launch}

\item{...}{other params}
}
\description{
launch boxplots_en
}
