% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{boxplots_fr}
\alias{boxplots_fr}
\title{Lance boxplots_fr}
\usage{
boxplots_fr(file = tutor:::tous_les_programmes()[10], ...)
}
\arguments{
\item{file}{name of the learnr to launch}

\item{...}{other params}
}
\description{
Lance boxplots_fr
}
