% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions_en.R
\name{customize_en}
\alias{customize_en}
\title{launch customize_en}
\usage{
customize_en(file = tutor:::tous_les_programmes(lang = "en")[16], ...)
}
\arguments{
\item{file}{name of the learnr to launch}

\item{...}{other params}
}
\description{
launch customize_en
}
