% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{customize_fr}
\alias{customize_fr}
\title{Lance customize_fr}
\usage{
customize_fr(file = tutor:::tous_les_programmes()[15], ...)
}
\arguments{
\item{file}{name of the learnr to launch}

\item{...}{other params}
}
\description{
Lance customize_fr
}
