% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions_en.R
\name{deriving_en}
\alias{deriving_en}
\title{launch deriving_en}
\usage{
deriving_en(file = tutor:::tous_les_programmes(lang = "en")[8], ...)
}
\arguments{
\item{file}{name of the learnr to launch}

\item{...}{other params}
}
\description{
launch deriving_en
}
