% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{deriving_fr}
\alias{deriving_fr}
\title{Lance deriving_fr}
\usage{
deriving_fr(file = tutor:::tous_les_programmes()[7], ...)
}
\arguments{
\item{file}{name of the learnr to launch}

\item{...}{other params}
}
\description{
Lance deriving_fr
}
