% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dxdy.R
\name{dxdy}
\alias{dxdy}
\title{Returns the nth derivative d^ny/dx^n}
\usage{
dxdy(x, y, n = 1)
}
\arguments{
\item{x}{x values}

\item{y}{y values}

\item{n}{order of the derivative}
}
\value{
a data.frame(x,y)
}
\description{
Returns the nth derivative d^ny/dx^n
}
\examples{
x <- seq(-10, 10, .0001)
y <- 1 + x + x^2 + x^3 + x^4
par(mfrow = c(2, 2), mar = c(4, 4, 1, 1))
plot(x, y, type = "l", main = "1+x+x^2+x^3+x^4")
plot(dxdy(x, y, 1), col = "red", type = "l", main = "1+2x+3x^2+4x^3")
plot(dxdy(x, y, 2), col = "blue", type = "l", main = "2+6x+12x^2")
plot(dxdy(x, y, 3), col = "orange", type = "l", main = "6+24x")
}
