% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions_en.R
\name{exploratory_data_analysis_en}
\alias{exploratory_data_analysis_en}
\title{launch exploratory_data_analysis_en}
\usage{
exploratory_data_analysis_en(
  file = tutor:::tous_les_programmes(lang = "en")[2],
  ...
)
}
\arguments{
\item{file}{name of the learnr to launch}

\item{...}{other params}
}
\description{
launch exploratory_data_analysis_en
}
