% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions_en.R
\name{histograms_en}
\alias{histograms_en}
\title{launch histograms_en}
\usage{
histograms_en(file = tutor:::tous_les_programmes(lang = "en")[9], ...)
}
\arguments{
\item{file}{name of the learnr to launch}

\item{...}{other params}
}
\description{
launch histograms_en
}
