% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{histograms_fr}
\alias{histograms_fr}
\title{Lance histograms_fr}
\usage{
histograms_fr(file = tutor:::tous_les_programmes()[8], ...)
}
\arguments{
\item{file}{name of the learnr to launch}

\item{...}{other params}
}
\description{
Lance histograms_fr
}
