% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integ.R
\name{integ}
\alias{integ}
\title{Simple numerical integration of y(x)}
\usage{
integ(x, y)
}
\arguments{
\item{x}{A vector containing x values}

\item{y}{A vector containing y values}
}
\value{

}
\description{
Simple numerical integration of y(x)
}
\examples{
x <- seq(0, 2 * pi, .0001)
y <- sin(x)^2
integ(x, y)
}
