% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions_en.R
\name{isolating_en}
\alias{isolating_en}
\title{launch isolating_en}
\usage{
isolating_en(file = tutor:::tous_les_programmes(lang = "en")[5], ...)
}
\arguments{
\item{file}{name of the learnr to launch}

\item{...}{other params}
}
\description{
launch isolating_en
}
