% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{isolating_fr}
\alias{isolating_fr}
\title{Lance isolating_fr}
\usage{
isolating_fr(file = tutor:::tous_les_programmes()[5], ...)
}
\arguments{
\item{file}{name of the learnr to launch}

\item{...}{other params}
}
\description{
Lance isolating_fr
}
