% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions_en.R
\name{join_datasets_en}
\alias{join_datasets_en}
\title{launch join_datasets_en}
\usage{
join_datasets_en(file = tutor:::tous_les_programmes(lang = "en")[10], ...)
}
\arguments{
\item{file}{name of the learnr to launch}

\item{...}{other params}
}
\description{
launch join_datasets_en
}
