% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{join_datasets_fr}
\alias{join_datasets_fr}
\title{Lance join_datasets_fr}
\usage{
join_datasets_fr(file = tutor:::tous_les_programmes()[9], ...)
}
\arguments{
\item{file}{name of the learnr to launch}

\item{...}{other params}
}
\description{
Lance join_datasets_fr
}
