% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{launch_learn}
\alias{launch_learn}
\title{Title}
\usage{
launch_learn(
  file = sample(tous_les_programmes(), 1),
  port = httpuv::randomPort(),
  host = "0.0.0.0"
)
}
\arguments{
\item{file}{file to launch}

\item{port}{port to use}

\item{host}{host}
}
\description{
Title
}
