% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions_en.R
\name{line_graphs_en}
\alias{line_graphs_en}
\title{launch line_graphs_en}
\usage{
line_graphs_en(file = tutor:::tous_les_programmes(lang = "en")[14], ...)
}
\arguments{
\item{file}{name of the learnr to launch}

\item{...}{other params}
}
\description{
launch line_graphs_en
}
