% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{line_graphs_fr}
\alias{line_graphs_fr}
\title{Lance line_graphs_fr}
\usage{
line_graphs_fr(file = tutor:::tous_les_programmes()[13], ...)
}
\arguments{
\item{file}{name of the learnr to launch}

\item{...}{other params}
}
\description{
Lance line_graphs_fr
}
