% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lorentzian.R
\name{Lorentzian}
\alias{Lorentzian}
\title{Sum of Lorentzian functions}
\usage{
Lorentzian(x, x0 = 0, FWHM = 1, A = 1, y0 = 0)
}
\arguments{
\item{x}{x values}

\item{x0}{Center values (scalar or vector)}

\item{FWHM}{Full width at half maximum (scalar or vector of length(x0))}

\item{A}{Amplitudes (scalar or vector of length(x0), defaults to '1')}

\item{y0}{baseline (scalar, defaults to '0')}
}
\value{

}
\description{
Sum of Lorentzian functions
}
\examples{
x <- seq(-10, 10, .1)
y <- Lorentzian(x, x0 = c(-2, 3), FW = c(1, 5), y0 = 1)
plot(x, y, type = "l")
}
