% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm01.R
\name{norm01}
\alias{norm01}
\title{Normalize a vector to [0,1]}
\usage{
norm01(x)
}
\arguments{
\item{x}{a vector to normalize}
}
\value{
(x-min(x))/(max(x)-min(x))
}
\description{
Normalize a vector to [0,1]
}
\examples{
x <- 1:10
norm01(x)
}
