% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions_en.R
\name{overplotting_en}
\alias{overplotting_en}
\title{launch overplotting_en}
\usage{
overplotting_en(file = tutor:::tous_les_programmes(lang = "en")[15], ...)
}
\arguments{
\item{file}{name of the learnr to launch}

\item{...}{other params}
}
\description{
launch overplotting_en
}
