% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{overplotting_fr}
\alias{overplotting_fr}
\title{Lance overplotting_fr}
\usage{
overplotting_fr(file = tutor:::tous_les_programmes()[14], ...)
}
\arguments{
\item{file}{name of the learnr to launch}

\item{...}{other params}
}
\description{
Lance overplotting_fr
}
