% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions_en.R
\name{programming_basics_en}
\alias{programming_basics_en}
\title{launch programming_basics_en}
\usage{
programming_basics_en(file = tutor:::tous_les_programmes(lang = "en")[12], ...)
}
\arguments{
\item{file}{name of the learnr to launch}

\item{...}{other params}
}
\description{
launch programming_basics_en
}
