% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{programming_basics_fr}
\alias{programming_basics_fr}
\title{Lance programming_basics_fr}
\usage{
programming_basics_fr(file = tutor:::tous_les_programmes()[11], ...)
}
\arguments{
\item{file}{name of the learnr to launch}

\item{...}{other params}
}
\description{
Lance programming_basics_fr
}
