% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pruby.R
\name{Pruby}
\alias{Pruby}
\title{Pressure <-> Ruby Raman shift \cr
Parameters from DOI: 10.1063/1.2135877}
\usage{
Pruby(w, laser = 532, P = -1, w0 = -1, l = FALSE)
}
\arguments{
\item{w}{Position of the R2 ruby line in cm-1}

\item{laser}{laser wavelength in nm (default=532)}

\item{P}{Pressure in GPa (default=-1). If P>0, will return the expected position of the R2 line in cm-1}

\item{w0}{Position of the R2 ruby line in cm-1 at P=0 (default=-1).\cr
If w0>0, corrects for potential strain in the ruby. Otherwise, uses the references lambda0 = 694.24 nm}

\item{l}{TRUE or FALSE (default). If TRUE and P>0, returns the expected position of the R2 line in nm instead of cm-1.}
}
\value{

}
\description{
Pressure <-> Ruby Raman shift \cr
Parameters from DOI: 10.1063/1.2135877
}
\examples{
Pruby(P = 1) # returns expected R2 position in cm-1
Pruby(w = 4401) # return corresponding pressure
}
