% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions_en.R
\name{reshape_data_en}
\alias{reshape_data_en}
\title{launch reshape_data_en}
\usage{
reshape_data_en(file = tutor:::tous_les_programmes(lang = "en")[3], ...)
}
\arguments{
\item{file}{name of the learnr to launch}

\item{...}{other params}
}
\description{
launch reshape_data_en
}
