% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rshift_to_lambda.R
\name{Rshift_to_lambda}
\alias{Rshift_to_lambda}
\title{Raman Shift -> Wavelength}
\usage{
Rshift_to_lambda(w, laser = 532)
}
\arguments{
\item{w}{Raman Shift in cm-1}

\item{laser}{laser wavelength in nm}
}
\value{
Returns the Raman Shift as a wavelength in nm
}
\description{
Raman Shift -> Wavelength
}
\examples{
Rshift_to_lambda(w = 4401)
}
