% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions_en.R
\name{scatterplots_en}
\alias{scatterplots_en}
\title{launch scatterplots_en}
\usage{
scatterplots_en(file = tutor:::tous_les_programmes(lang = "en")[13], ...)
}
\arguments{
\item{file}{name of the learnr to launch}

\item{...}{other params}
}
\description{
launch scatterplots_en
}
