% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{scatterplots_fr}
\alias{scatterplots_fr}
\title{Lance scatterplots_fr}
\usage{
scatterplots_fr(file = tutor:::tous_les_programmes()[12], ...)
}
\arguments{
\item{file}{name of the learnr to launch}

\item{...}{other params}
}
\description{
Lance scatterplots_fr
}
