% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions_en.R
\name{separate_columns_en}
\alias{separate_columns_en}
\title{launch separate_columns_en}
\usage{
separate_columns_en(file = tutor:::tous_les_programmes(lang = "en")[6], ...)
}
\arguments{
\item{file}{name of the learnr to launch}

\item{...}{other params}
}
\description{
launch separate_columns_en
}
