% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{separate_columns_fr}
\alias{separate_columns_fr}
\title{Lance separate_columns_fr}
\usage{
separate_columns_fr(file = tutor:::tous_les_programmes()[6], ...)
}
\arguments{
\item{file}{name of the learnr to launch}

\item{...}{other params}
}
\description{
Lance separate_columns_fr
}
