% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtraction.R
\name{subtraction}
\alias{subtraction}
\title{Subtract two vectors element-wise}
\usage{
subtraction(x, y)
}
\arguments{
\item{x}{First vector}

\item{y}{Second vector}
}
\value{
A vector containing the element-wise subtraction of y from x
}
\description{
Subtract two vectors element-wise
}
