% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions_en.R
\name{tibbles_en}
\alias{tibbles_en}
\title{launch tibbles_en}
\usage{
tibbles_en(file = tutor:::tous_les_programmes(lang = "en")[4], ...)
}
\arguments{
\item{file}{name of the learnr to launch}

\item{...}{other params}
}
\description{
launch tibbles_en
}
