% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{tibbles_fr}
\alias{tibbles_fr}
\title{Lance tibbles_fr}
\usage{
tibbles_fr(file = tutor:::tous_les_programmes()[3], ...)
}
\arguments{
\item{file}{name of the learnr to launch}

\item{...}{other params}
}
\description{
Lance tibbles_fr
}
