% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuto.R
\name{tuto}
\alias{tuto}
\title{tuto()}
\usage{
tuto(tutoname = NULL)
}
\arguments{
\item{tutoname}{Tutorial name. If NULL, prints the list of all available tutorials.}
}
\value{

}
\description{
Easy syntax to launch a tutorial.
}
