\name{tutoR-internal}
\title{Internal tutoR objects}
\alias{.abs}
\alias{.clean.html}
\alias{.clearScreen}
\alias{.Control}
\alias{.eg.html}
\alias{.exp}
\alias{.extra.examples}
\alias{.find.owner}
\alias{.function.exists}
\alias{.getexpr}
\alias{.get.varnames}
\alias{.is.num.expr}
\alias{.Log.checkX}
\alias{.Log}
\alias{.match.varname}
\alias{.Math}
\alias{.plot.checkXY}
\alias{.source}
\alias{.sqrt}
\alias{.Trig.checkX}
\alias{.Trig}
\alias{.var.exists}
\description{Internal tutoR objects.}
\details{These are not to be called by the user.}
\keyword{internal}