\name{tutoR.demo}
\alias{tutoR.demo}
\title{ Demonstration of the tutoR Package }
\description{
Demonstrates how function masks are used to pick up and
clarify bad inputs to the function before its execution.
New functions introduced, such as deskcheck() and eg()
are also demonstrated.
}
\usage{
tutoR.demo()
}
\seealso{
\code{\link{demo}}
}
\examples{
tutoR.demo()   # Or simply choose eg() Menu Option 1
}
\keyword{utilities}
