% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutoR_examples.R
\name{tutoR_examples}
\alias{tutoR_examples}
\title{Get path to tutoR example files}
\usage{
tutoR_examples(pattern = NULL)
}
\arguments{
\item{pattern}{Pattern to be found in file (regex filter). If `NULL`, no filter applied.}
}
\description{
tutoR comes bundled with a number of sample files in its `inst/extdata` directory. 
This function make them easy to access
}
\examples{
readr_example()
readr_example("challenge.csv")
}
