% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial.R
\name{tutorial}
\alias{tutorial}
\title{tuto}
\usage{
tutorial(tuto = NULL)
}
\arguments{
\item{tuto}{Tutorial name. If NULL, prints the list of all available tutorials.}
}
\value{

}
\description{
Easy syntax to launch a tutorial.
}
