% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions_en.R
\name{vis_basics_en}
\alias{vis_basics_en}
\title{launch vis_basics_en}
\usage{
vis_basics_en(file = tutor:::tous_les_programmes(lang = "en")[7], ...)
}
\arguments{
\item{file}{name of the learnr to launch}

\item{...}{other params}
}
\description{
launch vis_basics_en
}
