% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{exploratory_data_analysis_fr}
\alias{exploratory_data_analysis_fr}
\title{Lance exploratory_data_analysis_fr}
\usage{
exploratory_data_analysis_fr(file = tutor:::tous_les_programmes()[4], ...)
}
\arguments{
\item{file}{name of the learnr to launch}

\item{...}{other params}
}
\description{
Lance exploratory_data_analysis_fr
}
