% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{launch_learn}
\alias{launch_learn}
\title{Launch learnr}
\usage{
launch_learn(file = sample(tous_les_programmes(), 1), zoom = TRUE)
}
\arguments{
\item{file}{file to launch}

\item{zoom}{zoom on tutorial windows}
}
\description{
Launch learnr
}
