% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{launch_learn}
\alias{launch_learn}
\title{Launch learnr}
\usage{
launch_learn(
  file = sample(tous_les_programmes(), 1),
  zoom = TRUE,
  auto_kill_delay = 60 * 60 * 4,
  light = TRUE
)
}
\arguments{
\item{file}{file to launch}

\item{zoom}{zoom on tutorial windows}

\item{auto_kill_delay}{delay in secon to kill jobs}

\item{light}{light mode, boolean. if TRUE it goes direct to the point}
}
\description{
Launch learnr
}
