% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions_en.R
\name{programming_basics_en}
\alias{programming_basics_en}
\title{launch programming_basics_en}
\usage{
programming_basics_en(...)
}
\arguments{
\item{...}{other params}
}
\description{
launch programming_basics_en
}
