% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{programming_basics_fr}
\alias{programming_basics_fr}
\title{Lance programming_basics_fr}
\usage{
programming_basics_fr(...)
}
\arguments{
\item{...}{other params}
}
\description{
Lance programming_basics_fr
}
