% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions_en.R
\name{separate_columns_en}
\alias{separate_columns_en}
\title{launch separate_columns_en}
\usage{
separate_columns_en(...)
}
\arguments{
\item{...}{other params}
}
\description{
launch separate_columns_en
}
