\name{tutoR-internal}
\title{Internal tutoR objects}
\alias{.choose}
\alias{.clearScreen}
\alias{.Control}
\alias{.assist.help}
\alias{.assist.update}
\alias{.dim}
\alias{.Examples}
\alias{.error}
\alias{.execute}
\alias{.extra.examples}
\alias{.find.owner}
\alias{.function.exists}
\alias{.getexpr}
\alias{.get.varnames}
\alias{.is.numexpr}
\alias{.Log}
\alias{.match.varname}
\alias{.Math}
\alias{.menu}
\alias{.get.posint}
\alias{.plot.checkXY}
\alias{.Reserved}
\alias{.deletecall}
\alias{.setcall}
\alias{.Trig}
\alias{.var.exists}
\alias{.warning}
\description{Internal tutoR objects.}
\details{These are not to be called by the user.}
\keyword{internal}