\name{tutoR.demo}
\alias{tutoR.demo}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Demonstration of the student Package }
\description{
Demonstrates how function masks are used to pick up and
clarify bad inputs to the function before its execution.
New functions introduced, such as deskcheck() and eg()
are also demonstrated.
}
\usage{
tutoR.demo()
}
\examples{
tutoR.demo()          # Or choose Menu Option 1: eg()
}
\keyword{utilities}
