% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutor_examples.R
\name{tutor_examples}
\alias{tutor_examples}
\title{Get path to tutor example files}
\usage{
tutor_examples(pattern = NULL)
}
\arguments{
\item{pattern}{Pattern to be found in file (regex filter). If `NULL`, no filter applied.}
}
\description{
tutor comes bundled with a number of sample files in its `inst/extdata` directory. 
This function make them easy to access
}
\examples{
tutor_examples()
tutor_examples("people")
}
